+++ src/component/Customized.tsx
@@ -1,12 +1,12 @@
 /**
  * @fileOverview Customized
  */
-import React, { isValidElement, cloneElement, createElement, Component, FunctionComponent } from 'react';
+import React, { isValidElement, cloneElement, createElement, Component, FunctionComponent, ReactElement } from 'react';
 import _ from 'lodash';
 import { Layer } from '../container/Layer';
 import { warn } from '../util/LogUtils';
 
-type Comp<P> = FunctionComponent<P> | Component<P>;
+type Comp<P> = FunctionComponent<P> | Component<P> | ReactElement<P>;
 export type Props<P, C extends Comp<P>> = P & {
   component: C;
 };